move hex colours to theme level
