+++ editor.planx.uk/src/pages/FlowEditor/ReadMePage/ReadMePage.tsx
@@ -7,7 +7,8 @@ import { useFormik } from "formik";
 import { useToast } from "hooks/useToast";
 import capitalize from "lodash/capitalize.js";
 import React from "react";
-import FlowTag, { FlowTagType, StatusVariant } from "ui/editor/FlowTag";
+import FlowTag from "ui/editor/FlowTag/FlowTag";
+import { FlowTagType, StatusVariant } from "ui/editor/FlowTag/types";
 import InputGroup from "ui/editor/InputGroup";
 import InputLegend from "ui/editor/InputLegend";
 import RichTextInput from "ui/editor/RichTextInput/RichTextInput";
