+++ editor.planx.uk/src/ui/editor/FlowTag/FlowTag.tsx
@@ -3,31 +3,11 @@ import { styled } from "@mui/material/styles";
 import React from "react";
 import { FONT_WEIGHT_SEMI_BOLD } from "theme";
 
-export const FlowTagType = {
-  Status: "status",
-  ApplicationType: "applicationType",
-  ServiceType: "serviceType",
-} as const;
-
-type ObjectValues<T> = T[keyof T];
-
-type FlowTagType = ObjectValues<typeof FlowTagType>;
-
-export const StatusVariant = {
-  Online: "online",
-  Offline: "offline",
-} as const;
-
-type StatusVariant = ObjectValues<typeof StatusVariant>;
-
-const BG_ONLINE = "#D6FFD7";
-const BG_OFFLINE = "#EAEAEA";
-const BG_APPLICATION_TYPE = "#D6EFFF";
-const BG_SERVICE_TYPE = "#FFEABE";
+import { FlowTagProps, FlowTagType, StatusVariant } from "./types";
 
 const Root = styled(Box, {
   shouldForwardProp: (prop) => prop !== "tagType" && prop !== "statusVariant",
-})<Props>(({ theme, tagType, statusVariant }) => ({
+})<FlowTagProps>(({ theme, tagType, statusVariant }) => ({
   fontSize: theme.typography.body2.fontSize,
   fontWeight: FONT_WEIGHT_SEMI_BOLD,
   padding: "2px 6px",
@@ -39,7 +19,9 @@ const Root = styled(Box, {
   border: "1px solid rgba(0, 0, 0, 0.2)",
   ...(tagType === FlowTagType.Status && {
     backgroundColor:
-      statusVariant === StatusVariant.Online ? BG_ONLINE : BG_OFFLINE,
+      statusVariant === StatusVariant.Online
+        ? theme.palette.flowTag.online
+        : theme.palette.flowTag.offline,
     "&::before": {
       content: '""',
       width: "8px",
@@ -52,21 +34,19 @@ const Root = styled(Box, {
     },
   }),
   ...(tagType === FlowTagType.ApplicationType && {
-    backgroundColor: BG_APPLICATION_TYPE,
+    backgroundColor: theme.palette.flowTag.applicationType,
   }),
   ...(tagType === FlowTagType.ServiceType && {
-    backgroundColor: BG_SERVICE_TYPE,
+    backgroundColor: theme.palette.flowTag.serviceType,
   }),
 }));
 
-export interface Props {
-  id?: string;
-  tagType: FlowTagType;
-  statusVariant?: StatusVariant;
-  children?: React.ReactNode;
-}
-
-const FlowTag: React.FC<Props> = ({ id, tagType, statusVariant, children }) => (
+const FlowTag: React.FC<FlowTagProps> = ({
+  id,
+  tagType,
+  statusVariant,
+  children,
+}) => (
   <Root id={id} tagType={tagType} statusVariant={statusVariant}>
     {children}
   </Root>
